Integrating Dynamic Reconfigure for the configuration.
For now, only implemented PID to test it (with the config_test.cc)

Configuration are changeable only via the dynamic reconfigure for now.
When we load the node, we read the config file and set the configuraiotn to the dynamic reconf.
When we set a configuration, we save the file directly, so configuraiton are always up to date.

Each configuration manager should inherit from config_manager.h.

Adding property.h to hold constant n stuff