Adding configuration handler for PID, Node and Thruster.
Those configuration are not final, they are only for testing and basics.

the configuration manager read a configuration, upload a dynamic reconfigure server. when the configuration
change, it saves the change to file. Upon loading (at the beginnig of the program) it sets the defaut to the current
configuration.

For now the members are private (for the config data). The person coding the usage can modify it the way he/she wants it.