Big commit before refactor.
*** Please do not follow this example of big commit... ***

Managed to make the dynamic reconfigure work and the signal/slot work too.

Adding the Thruster control and the 4axis PID algorithm

Modification to NodeConfig for AlgorithmConfig.

Adding the configuration for PID config

Adding bunch of manager that will be reduced in next commit  (why have a config manager AND a manager that holds a config manager?)

Adding services for setting target and enabling directions